update db int->bigint

+++ bot/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, ForeignKey, TIMESTAMP, func
+from sqlalchemy import Column,String, Boolean, ForeignKey, TIMESTAMP, func, BigInteger
 from sqlalchemy import DateTime
 from sqlalchemy.orm import relationship
 from .base import Base
@@ -13,7 +13,7 @@ ban_expiration = Column(DateTime(timezone=True))
 
 class CardType(Base):
     __tablename__ = "card_types"
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     first_name = Column(String(100), nullable=False)
     last_name = Column(String(100), nullable=False)
 
@@ -23,7 +23,7 @@ class CardType(Base):
 
 class Collection(Base):
     __tablename__ = "collections"
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     slug = Column(String(100), unique=True, nullable=False)
     title = Column(String(150), nullable=False)
     is_primary = Column(Boolean, default=False)
@@ -33,8 +33,8 @@ class Collection(Base):
 
 class CardTypeInCollection(Base):
     __tablename__ = "cardtype_collections"
-    card_type_id = Column(Integer, ForeignKey("card_types.id"), primary_key=True)
-    collection_id = Column(Integer, ForeignKey("collections.id"), primary_key=True)
+    card_type_id = Column(BigInteger, ForeignKey("card_types.id"), primary_key=True)
+    collection_id = Column(BigInteger, ForeignKey("collections.id"), primary_key=True)
     image_path = Column(String(255), nullable=True)
 
     card_type = relationship("CardType", back_populates="memberships")
@@ -45,7 +45,7 @@ class CardTypeInCollection(Base):
 
 class User(Base):
     __tablename__ = "users"
-    id = Column(Integer, primary_key=True, index=True)  # Telegram user id
+    id = Column(BigInteger, primary_key=True, index=True)  # Telegram user id
     username = Column(String(100))
     first_name = Column(String(100))
     last_name = Column(String(100))
@@ -55,7 +55,7 @@ class User(Base):
     is_banned = Column(Boolean, default=False)
     ban_expiration = Column(TIMESTAMP)
 
-    failed_attempts = Column(Integer, default=0)
+    failed_attempts = Column(BigInteger, default=0)
     last_failed_at = Column(TIMESTAMP)
     ban_until = Column(TIMESTAMP)
 
@@ -65,12 +65,12 @@ class User(Base):
 
 class Card(Base):
     __tablename__ = "cards"
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     code = Column(String(50), unique=True, nullable=False)
 
-    card_type_id = Column(Integer, ForeignKey("card_types.id"), nullable=False)
-    prize_id = Column(Integer, ForeignKey("prizes.id"))
-    activated_by = Column(Integer, ForeignKey("users.id"))
+    card_type_id = Column(BigInteger, ForeignKey("card_types.id"), nullable=False)
+    prize_id = Column(BigInteger, ForeignKey("prizes.id"))
+    activated_by = Column(BigInteger, ForeignKey("users.id"))
     activated_at = Column(TIMESTAMP)
     is_activated = Column(Boolean, default=False)
 
@@ -81,7 +81,7 @@ class Card(Base):
 
 class Prize(Base):
     __tablename__ = "prizes"
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     title = Column(String(100), nullable=False)
     description = Column(String)
     cards = relationship("Card", back_populates="prize")
